requirements.txt
braintree==3.57.1

certifi==2019.11.28
chardet==3.0.4
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==2.2.1
django-ckeditor==5.7.0
django-crispy-forms==1.8.0
django-js-asset==1.2.2
django-payments==0.13.0
django-taggit==1.1.0
idna==2.8
Pillow==6.0.0
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
six==1.13.0
sqlparse==0.3.0
stripe==2.38.0
suds-jurko==0.6
urllib3==1.25.6
xmltodict==0.12.0
